writing main
